+++ meson.build
@@ -21,4 +21,11 @@ py_mod = import('python')
 py = py_mod.find_installation()
 py_dep = py.dependency()
 
+py.install_sources([
+    'version.py',
+],
+  pure: false,    # Will be installed next to binaries
+  subdir: 'freesas'  # Folder relative to site-packages to install to
+)
+
 subdir('src/freesas')
